+++ libs/components/code-examples/src/lib/modules/ag-grid/data-grid/data-manager-multiselect/example.component.ts
@@ -7,36 +7,48 @@ import {
   inject,
 } from '@angular/core';
 import {
-  SkyDataManagerConfig,
   SkyDataManagerModule,
   SkyDataManagerService,
   SkyDataManagerState,
 } from '@skyux/data-manager';
+import {
+  SkyFilterBarModule,
+  SkyFilterItemLookupSearchAsyncArgs,
+} from '@skyux/filter-bar';
+import { SkyListSummaryModule } from '@skyux/lists';
 
 import { Subject, takeUntil } from 'rxjs';
 
 import { AG_GRID_DEMO_DATA } from './data';
-import { FilterModalComponent } from './filter-modal.component';
-import { Filters } from './filters';
+import { ExampleService } from './example.service';
+import { SalesModalComponent } from './sales-modal.component';
 import { ViewGridComponent } from './view-grid.component';
 
 /**
- * @title Data manager multiselect setup
+ * @title Data manager setup
  */
 @Component({
-  selector: 'app-ag-grid-data-grid-data-manager-multiselect-example',
+  selector: 'app-ag-grid-data-grid-data-manager-example',
   templateUrl: './example.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
   providers: [SkyDataManagerService],
-  imports: [SkyDataManagerModule, ViewGridComponent],
+  imports: [
+    SkyDataManagerModule,
+    SkyFilterBarModule,
+    SkyListSummaryModule,
+    ViewGridComponent,
+  ],
 })
 export class AgGridDataGridDataManagerMultiselectExampleComponent
   implements OnInit, OnDestroy
 {
   protected items = AG_GRID_DEMO_DATA;
 
-  #dataManagerConfig: SkyDataManagerConfig = {
-    filterModalComponent: FilterModalComponent,
+  protected salesModal = SalesModalComponent;
+
+  #activeViewId = 'dataGridWithDataManagerView';
+
+  #dataManagerConfig = {
     sortOptions: [
       {
         id: 'az',
@@ -55,17 +67,12 @@ export class AgGridDataGridDataManagerMultiselectExampleComponent
 
   #defaultDataState = new SkyDataManagerState({
     filterData: {
-      filtersApplied: false,
-      filters: {
-        hideSales: false,
-        jobTitle: 'any',
-      } satisfies Filters,
+      filters: {},
     },
     views: [
       {
-        viewId: 'dataGridMultiselectWithDataManagerView',
+        viewId: 'dataGridWithDataManagerView',
         displayedColumnIds: [
-          'selected',
           'context',
           'name',
           'age',
@@ -78,11 +85,11 @@ export class AgGridDataGridDataManagerMultiselectExampleComponent
     ],
   });
 
-  #activeViewId = 'dataGridMultiselectWithDataManagerView';
   #ngUnsubscribe = new Subject<void>();
 
   readonly #changeDetectorRef = inject(ChangeDetectorRef);
   readonly #dataManagerSvc = inject(SkyDataManagerService);
+  readonly #exampleSvc = inject(ExampleService);
 
   constructor() {
     this.#dataManagerSvc
@@ -105,5 +112,13 @@ export class AgGridDataGridDataManagerMultiselectExampleComponent
   public ngOnDestroy(): void {
     this.#ngUnsubscribe.next();
     this.#ngUnsubscribe.complete();
+  }
+
+  public onJobTitleSearchAsync(args: SkyFilterItemLookupSearchAsyncArgs): void {
+    // In a real-world application the search service might return an Observable
+    // created by calling HttpClient.get(). Assigning that Observable to the result
+    // allows the lookup component to cancel the web request if it does not complete
+    // before the user searches again.
+    args.result = this.#exampleSvc.search(args.searchText);
   }
 }